Fix more warnings.
Check in updated web page.



git-svn-id: http://svn.code.sf.net/p/genoviz/code/trunk@3858 80d5a9c2-7d8c-4c9f-9369-4e03bca4d2bd

+++ igb/src/com/affymetrix/igb/glyph/ThreshGlyph.java
@@ -39,7 +39,7 @@ public final class ThreshGlyph extends FillRectGlyph {
     view.transformToPixels(coordbox, pixelbox);
     Graphics g = view.getGraphics();
     g.setColor(this.getBackgroundColor());
-    int xpos = (int)Math.max(pixelbox.x, 0);
+    int xpos = Math.max(pixelbox.x, 0);
     g.drawString(label, xpos+20, pixelbox.y-2);
   }
 }
